허브경로 단건조회

+++ com.msa-delivery.hub/src/main/java/com/msa_delivery/hub/presentation/response/HubRouteResponse.java
@@ -14,8 +14,8 @@ public record HubRouteResponse(
         UUID hubRouteId,
         UUID departureHubId,
         UUID arrivalHubId,
-        double distance,
-        double duration
+        Integer distance,
+        Integer duration
 ) {
     public static HubRouteResponse from(HubRoute hubRoute) {
         return HubRouteResponse.builder()
